WIP removing windows dependency

+++ iModelBridgeCore/Imagepp/PublicApi/ImagePP/all/h/HMGThread.h
@@ -15,7 +15,7 @@
 
 BEGIN_IMAGEPP_NAMESPACE
 
-typedef int HMGThreadID;
+typedef intptr_t HMGThreadID;
 
 class HMGSink;
 class HMGMessage;
